requirements.txt
aiohttp
aiomysql
pydantic
google-generativeai
langchain
chromadb
huggingface-hub
sentence-transformers
python-dotenv